+++ BL.API/BL.API.Services/Stats/Model/PlayerByFactionWRItem.cs
@@ -4,14 +4,14 @@
     {
         public string Nickname { get; set; }
         public int AseraiCount { get; set; }
-        public decimal AseraiWR { get; set; }
+        public double AseraiWR { get; set; }
         public int BattaniaCount { get; set; }
-        public decimal BattaniaWR { get; set; }
+        public double BattaniaWR { get; set; }
         public int EmpireCount { get; set; }
-        public decimal EmpireWR { get; set; }
+        public double EmpireWR { get; set; }
         public int KhuzaitCount { get; set; }
-        public decimal KhuzaitWR { get; set; }
+        public double KhuzaitWR { get; set; }
         public int VlandiaCount { get; set; }
-        public decimal VlandiaWR { get; set; }
+        public double VlandiaWR { get; set; }
     }
 }
